SaaStelligence/requirements.txt
# Optional integrations. Core training and agent logic run on the Python standard library.
langchain==0.0.338
requests==2.31.0
python-dotenv==1.0.0
gspread==5.9.0
oauth2client==4.1.3
